Moved setbookingproperties to updateBookingHandler
