+++ src/main/java/com/wirelessiths/dal/Booking.java
@@ -260,53 +260,4 @@ public class Booking {
         return true;
     }
 
-    /**
-     * Makes a booking object from updateBookingRequest. Should be used for updating object so that if no new info is inputted, the old information will persist.
-     * @param updateBookingRequest - the booking request in where the new information is stored.
-     * @param booking - the booking that you want to update.
-     * @return updated booking
-     */
-   public static Booking setBookingProperties(UpdateBookingRequest updateBookingRequest, Booking booking) {
-
-        updateBookingRequest.getUserId().ifPresent(booking::setUserId);
-        updateBookingRequest.getScooterId().ifPresent(booking::setScooterId);
-        updateBookingRequest.getBookingId().ifPresent(booking::setBookingId);
-
-        updateBookingRequest.getDate().ifPresent(n -> {
-            try {
-                LocalDateConverter converter = new LocalDateConverter();
-                booking.setDate(converter.unconvert(n));
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        });
-
-        updateBookingRequest.getStartTime().ifPresent(n -> {
-            try {
-                InstantConverter converter = new InstantConverter();
-                booking.setStartTime(converter.unconvert(n));
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        });
-
-        updateBookingRequest.getEndTime().ifPresent(n -> {
-            try {
-                InstantConverter converter = new InstantConverter();
-                booking.setEndTime(converter.unconvert(n));
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        });
-
-       updateBookingRequest.getTripStatus().ifPresent(n -> {
-           if(n.equals("WAITING_TO_START") || n.equals("IN_PROGRESS") || n.equals("COMPLETED") || n.equals("SCOOTER_NOT_RETURNED")) {
-               TripStatus tripStatus = TripStatus.valueOf(n);
-               booking.setTripStatus(tripStatus);
-           }
-
-        });
-
-        return booking;
-    }
 }
